@@ -30,43 +30,43 @@
   },
   "author": {
     "name": "Philipp Garbe",
-    "url": "https://garbe.io",
+    "url": "https://github.com/pgarbe",
     "organization": false
   },
   "devDependencies": {
-    "@aws-cdk/assert": "^2.0.0-rc.29",
-    "@types/jest": "^26.0.24",
+    "@aws-cdk/assert": "2.0.0",
+    "@types/jest": "^27.0.3",
     "@types/node": "^14.15.0",
-    "@typescript-eslint/eslint-plugin": "^4.31.0",
-    "@typescript-eslint/parser": "^4.31.0",
-    "aws-cdk-lib": "2.0.0-rc.29",
+    "@typescript-eslint/eslint-plugin": "^5",
+    "@typescript-eslint/parser": "^5",
+    "aws-cdk-lib": "2.0.0",
     "constructs": "10.0.5",
-    "esbuild": "^0.12.25",
-    "eslint": "^7.32.0",
+    "esbuild": "^0.14.2",
+    "eslint": "^8",
     "eslint-import-resolver-node": "^0.3.6",
-    "eslint-import-resolver-typescript": "^2.4.0",
-    "eslint-plugin-import": "^2.24.2",
-    "jest": "^26.6.3",
-    "jest-junit": "^12",
-    "jsii": "^1.34.0",
-    "jsii-diff": "^1.34.0",
-    "jsii-docgen": "^1.8.110",
-    "jsii-pacmak": "^1.34.0",
-    "json-schema": "^0.3.0",
+    "eslint-import-resolver-typescript": "^2.5.0",
+    "eslint-plugin-import": "^2.25.3",
+    "jest": "^27.4.3",
+    "jest-junit": "^13",
+    "jsii": "^1.46.0",
+    "jsii-diff": "^1.46.0",
+    "jsii-docgen": "^4.2.6",
+    "jsii-pacmak": "^1.46.0",
+    "json-schema": "^0.4.0",
     "npm-check-updates": "^11",
     "pre-commit": "^1.2.2",
-    "projen": "^0.33.4",
+    "projen": "^0.37.1",
     "standard-version": "^9",
-    "ts-jest": "^26.5.6",
-    "typescript": "^4.4.2"
+    "ts-jest": "^27.1.0",
+    "ts-node": "^9",
+    "typescript": "^4.5.2"
   },
   "peerDependencies": {
-    "aws-cdk-lib": "^2.0.0-rc.29",
+    "aws-cdk-lib": "2.0.0",
     "constructs": "^10.0.5"
   },
   "dependencies": {
-    "aws-cdk-lib": "^2.0.0-rc.29",
-    "aws-sdk": "^2.985.0",
+    "aws-sdk": "^2.1043.0",
     "jszip": "^3.7.1"
   },
   "bundledDependencies": [
