@@ -1,6 +1,6 @@
 import * as docker from '../src/lambda/docker-adapter';
 
-test('Docker image tags are loaded through pagination', async (done) => {
+test('Docker image tags are loaded through pagination', async () => {
 
   // WHEN
   let tags = await docker.getDockerImageTags('datadog/agent');
@@ -10,11 +10,9 @@ test('Docker image tags are loaded through pagination', async (done) => {
   expect(tags[0].digest?.length).toBeGreaterThan(0);
   expect(tags[0].tag.length).toBeGreaterThan(0);
 
-  done();
-
 }, 30000);
 
-test('Docker library image tags are loaded through pagination', async (done) => {
+test('Docker library image tags are loaded through pagination', async () => {
 
   // WHEN
   let tags = await docker.getDockerImageTags('amazonlinux');
@@ -24,20 +22,16 @@ test('Docker library image tags are loaded through pagination', async (done) => 
   expect(tags[0].digest?.length).toBeGreaterThan(0);
   expect(tags[0].tag.length).toBeGreaterThan(0);
 
-  done();
-
 }, 30000);
 
 
-test('Docker library image tags include only amd64/linux images', async (done) => {
+test('Docker library image tags include only amd64/linux images', async () => {
 
   // WHEN
   let tags = await docker.getDockerImageTags('mongo');
 
   // THEN
   expect(tags.filter(t => t.tag === '4.2.4').length).toBe(0); // mongo:4.2.4 has only amd64/windows images
-
-  done();
 
 }, 30000);
 
